refactor(assistant): move the companion surface controller into electron-desktop
The controller, its capture sources and the pointer flag move out of
clients/macos/src/main into packages/electron-desktop/src, behind a
CompanionPlatform seam for the IPC registrars, the app's window and
reduced motion, and a CompanionCapturePlatform seam for the capture
backend and the Chrome-tab step. clients/macos/src/main/companion-window.ts
stays as a thin adapter that wires the macOS pieces, so every importer
and the existing tests are unchanged.
