+++ packages/electron-desktop/package.json
@@ -21,7 +21,10 @@
     "./capability-registry": "./src/capability-registry.ts",
     "./command-palette-window": "./src/command-palette-window.ts",
     "./commands": "./src/commands.ts",
+    "./companion-capture-sources": "./src/companion-capture-sources.ts",
     "./companion-menu": "./src/companion-menu.ts",
+    "./companion-pointer": "./src/companion-pointer.ts",
+    "./companion-window": "./src/companion-window.ts",
     "./connectivity-probe": "./src/connectivity-probe.ts",
     "./csp": "./src/csp.ts",
     "./deep-links": "./src/deep-links.ts",
