+++ packages/electron-desktop/src/companion-pointer.ts
